Rename to snake case and remove UB
